requirements.txt
pyspark==3.5.3
kafka-python==2.0.2
apache-airflow==2.8.4
yfinance==0.2.36
requests==2.31.0
pydantic==2.6.4
pydantic-settings==2.2.1
boto3==1.34.62
streamlit==1.32.2
pandas==2.2.1
pyarrow==15.0.2
